Use zero-copy for ramcloud and noop backends
Changed the externram API to pass double pointers to write, read,
read_top, read_bottom. To handle flexibility between backends that
support zero-copy, new return values are added to write and multiWrite.
This allows libuserfault to known when it is responsible for releasing a
page buffer.

For writes: page ownership is transferred to externram, which is free to
use the given buffer to insert the page into its key-value store. It is
expected that externram will communicate back to libuserfault on what
should be done with the buffer it was given. For memcached, the buffer
will need to be freed by libuserfault.

For reads: libuserfault or libpagecache will not be expected to allocate
a buffer before calling the externram function. Page buffers are reused.
If the backend does not support zero-copy (e.g. memcached), then a
buffer will need to be allocated in multiRead

In order to get threaded prefetch to work with asyn read, it was
necessary to establish mutual exclusivity when reading from externram.
The prefetch thread should not attempt to read while the userfault
handler is actively reading a page. It should wait until userfault
handler begins polling.

Change-Id: I090ad3d36421f47cd53e2a174e6c8df2464c1aba