Keep trying to evict encountered zeropages
When the uffd eviction returns EBUSY, the page is a zeropage (hasn't
been written to by the client). We can't throw it away, because the
guest might write to it, which wouldn't cause a uffd event. Our only
option is to hold on to it and put it back on the LRU.

When the size of the LRU exceeds capacity, try to evict pages after
having returned the faulted page to the application, so it is not
blocked waiting for its page fault to be handled.

Change-Id: Ie911c48105e19d1c3410e6a8f871dd4ce8574764